+++ app/models/group.rb
@@ -6,8 +6,8 @@ class Group < ActiveRecord::Base
 
   has_many :group_memberships
   has_many :group_submissions
-  has_many :projects, -> { where("group_memberships.active = :value and projects.enrolled = true", value: true) }, through: :group_memberships
-  has_many :past_projects, -> { where("group_memberships.active = :value", value: false) },  through: :group_memberships, source: 'project'
+  has_many :projects, -> { where('group_memberships.active = :value and projects.enrolled = true', value: true) }, through: :group_memberships
+  has_many :past_projects, -> { where('group_memberships.active = :value', value: false) }, through: :group_memberships, source: 'project'
   has_one :unit, through: :group_set
   has_one :tutor, through: :tutorial
 
@@ -16,7 +16,7 @@ class Group < ActiveRecord::Base
   validates :tutorial, presence: true, allow_nil: false
   validates_associated :group_memberships
   validates :name, uniqueness: { scope: :group_set,
-    message: "must be unique within the set of groups" }
+                                 message: 'must be unique within the set of groups' }
   validate :must_be_in_same_tutorial, if: :limit_members_to_tutorial?
 
   before_destroy :ensure_no_submissions
@@ -46,26 +46,24 @@ class Group < ActiveRecord::Base
 
     # Return permissions hash
     {
-      :convenor => convenor_role_permissions,
-      :tutor    => tutor_role_permissions,
-      :student  => student_role_permissions,
-      :nil      => nil_role_permissions
+      convenor: convenor_role_permissions,
+      tutor: tutor_role_permissions,
+      student: student_role_permissions,
+      nil: nil_role_permissions
     }
   end
 
-  def ensure_no_submissions()
-    return true if group_submissions.count == 0
-    self.errors[:base] << "Cannot delete group while it has submissions."
-    return false
+  def ensure_no_submissions
+    return true if group_submissions.count.zero?
+    errors[:base] << 'Cannot delete group while it has submissions.'
+    false
   end
 
-  def specific_permission_hash(role, perm_hash, other)
+  def specific_permission_hash(role, perm_hash, _other)
     result = perm_hash[role] unless perm_hash.nil?
-  if result && role == :student
-      if group_set.allow_students_to_manage_groups
-        result << :manage_group
+    if result && role == :student
+      result << :manage_group if group_set.allow_students_to_manage_groups
       end
-    end
     result
   end
 
@@ -79,14 +77,14 @@ class Group < ActiveRecord::Base
   end
 
   def has_user(user)
-    projects.where("user_id = :user_id", user_id: user.id).count == 1
+    projects.where('user_id = :user_id', user_id: user.id).count == 1
   end
 
   def add_member(project)
     gm = project.group_membership_for_groupset(group_set)
 
     if gm.nil?
-      gm = GroupMembership.create(group: self, project:project)
+      gm = GroupMembership.create(group: self, project: project)
     else
       gm = GroupMembership.find(gm.id)
       gm.group = self
@@ -108,15 +106,15 @@ class Group < ActiveRecord::Base
   #
   # check if the project is the same as the current submission
   #
-  def __different_project_composition__ (contributors, gs)
-    logger.debug "Starting checks"
+  def __different_project_composition__(contributors, gs)
+    logger.debug 'Starting checks'
     contributors.each do |contrib|
       logger.debug "-- Checking #{contrib}"
       return true unless gs.projects.include? contrib[:project]
       return true unless contrib[:pct].to_i > 0
     end
     logger.debug "Checking #{contributors.count} == #{gs.projects.count}"
-    return contributors.count != gs.projects.count
+    contributors.count != gs.projects.count
   end
 
   #
@@ -128,7 +126,7 @@ class Group < ActiveRecord::Base
   #
   def create_submission(submitter_task, notes, contributors)
     total = 0
-    #check all members are in the same group
+    # check all members are in the same group
     contributors.each do |contrib|
       project = contrib[:project]
       pct = contrib[:pct].to_i
@@ -143,7 +141,7 @@ class Group < ActiveRecord::Base
       contrib[:pts] = 0 unless pts > 0
       contrib[:pts] = 5 unless pts < 5
 
-      raise "Not all contributions were from team members." unless projects.include? project
+      raise 'Not all contributions were from team members.' unless projects.include? project
     end
 
     # check for all group members
@@ -154,13 +152,13 @@ class Group < ActiveRecord::Base
     raise 'Contribution percentages are excessive.' unless total <= 110
 
     # check group task
-    raise "Group submission only allowed for group tasks." unless submitter_task.task_definition.group_set
-    raise "Group submission for wrong group for unit." unless submitter_task.task_definition.group_set == group_set
+    raise 'Group submission only allowed for group tasks.' unless submitter_task.task_definition.group_set
+    raise 'Group submission for wrong group for unit.' unless submitter_task.task_definition.group_set == group_set
 
     old_gs = submitter_task.group_submission
     gs = old_gs
     if gs.nil? || __different_project_composition__(contributors, gs)
-      gs = GroupSubmission.create()
+      gs = GroupSubmission.create
       gs.task_definition = submitter_task.task_definition
     end
 
@@ -183,12 +181,10 @@ class Group < ActiveRecord::Base
 
     if old_gs
       old_gs.reload
-      if old_gs.projects.count == 0
-        old_gs.destroy!
-      end
+      old_gs.destroy! if old_gs.projects.count.zero?
     end
 
-    #ensure that original task is reloaded... update will have effected a different object
+    # ensure that original task is reloaded... update will have effected a different object
     submitter_task.reload
     gs
   end
@@ -199,7 +195,7 @@ class Group < ActiveRecord::Base
 
   def must_be_in_same_tutorial
     if limit_members_to_tutorial?
-      if ! all_members_in_tutorial?
+      unless all_members_in_tutorial?
         errors.add(:members, "must all be in the group's tutorial")
       end
     end
@@ -210,8 +206,8 @@ class Group < ActiveRecord::Base
   #
   def all_members_in_tutorial?
     group_memberships.each do |member|
-      return false unless (not member.active) || member.in_group_tutorial?(self.tutorial)
+      return false unless !member.active || member.in_group_tutorial?(tutorial)
     end
-    return true
+    true
   end
 end
